luci-base: form.js: Fix CBITypedSection sections adding/removing
Add or remove sections only if map.save() completes successfully.

If we add or remove a sections before calling map.save(), then
if the form contains errors, we will get a new added section, but
we will not see it in the form (no rendering performed). Then,
when we fix the errors in the form and try to add the new section
again, we can get several new sections (old, which are not
rendered due to form validation errors and new) instead of one.

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>